backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
sqlalchemy==2.0.36
scikit-learn==1.5.2
numpy==1.26.4
python-dotenv==1.0.1
eventlet==0.37.0
